# openapi/kit-merged.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Realtime Service",
    "version": "1.0"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/auth/web/consent": {
      "get": {
        "operationId": "AuthWeb_GetConsent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetConsentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "consentChallenge",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "AuthWeb"
        ]
      },
      "post": {
        "operationId": "AuthWeb_GrantConsent",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GrantConsentResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GrantConsentRequest"
            }
          }
        ],
        "tags": [
          "AuthWeb"
        ]
      }
    },
    "/v1/auth/web/login": {
      "get": {
        "operationId": "AuthWeb_GetWebLogin",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetLoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "redirect",
            "description": "redirect url.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "loginChallenge",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "AuthWeb"
        ]
      },
      "post": {
        "operationId": "AuthWeb_WebLogin",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1WebLoginAuthReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1WebLoginAuthRequest"
            }
          }
        ],
        "tags": [
          "AuthWeb"
        ]
      }
    },
    "/v1/auth/web/logout": {
      "get": {
        "operationId": "AuthWeb_GetWebLogout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetLogoutResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "logoutChallenge",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "AuthWeb"
        ]
      },
      "post": {
        "operationId": "AuthWeb_WebLogout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LogoutResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LogoutRequest"
            }
          }
        ],
        "tags": [
          "AuthWeb"
        ]
      }
    },
...
  },
  "definitions": {
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "v1GetConsentResponse": {
      "type": "object",
      "properties": {
        "challenge": {
          "type": "string",
          "required": [
            "challenge"
          ]
        },
        "requestedScope": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "userId": {
          "type": "string"
        },
        "client": {
          "$ref": "#/definitions/v1OAuthClient"
        },
        "redirect": {
          "type": "string"
        }
      },
      "required": [
        "challenge"
      ]
    },
...
    "v1GetLoginResponse": {
      "type": "object",
      "properties": {
        "redirect": {
          "type": "string",
          "title": "normalized redirect url"
        },
        "oauth": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1OAuthProvider"
          }
        },
        "hint": {
          "type": "string"
        },
        "challenge": {
          "type": "string"
        }
      }
    },
    "v1GetLogoutResponse": {
      "type": "object",
      "properties": {
        "challenge": {
          "type": "string"
        }
      }
    },
...
    "v1GrantConsentRequest": {
      "type": "object",
      "properties": {
        "challenge": {
          "type": "string"
        },
        "grantScope": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "user granted scope"
        },
        "reject": {
          "type": "boolean"
        }
      }
    },
    "v1GrantConsentResponse": {
      "type": "object",
      "properties": {
        "redirect": {
          "type": "string"
        }
      }
    },
...
    "v1LogoutRequest": {
      "type": "object",
      "properties": {
        "challenge": {
          "type": "string"
        }
      }
    },
    "v1LogoutResponse": {
      "type": "object",
      "properties": {
        "redirect": {
          "type": "string"
        }
      }
    },
...
    "v1OAuthClient": {
      "type": "object",
      "properties": {
        "clientId": {
          "type": "string",
          "x-nullable": true
        },
        "clientName": {
          "type": "string",
          "x-nullable": true
        },
        "clientUri": {
          "type": "string",
          "x-nullable": true
        },
        "contacts": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "logoUri": {
          "type": "string",
          "x-nullable": true
        },
        "metadata": {
          "type": "object",
          "x-nullable": true
        },
        "owner": {
          "type": "string",
          "x-nullable": true
        },
        "policyUri": {
          "type": "string",
          "x-nullable": true
        }
      }
    },
    "v1OAuthProvider": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
...
    "v1WebLoginAuthReply": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "tokenType": {
          "type": "string",
          "title": "Bearer"
        },
        "expires": {
          "type": "integer",
          "format": "int32",
          "title": "Expires time in seconds"
        },
        "refreshToken": {
          "type": "string",
          "title": "Refresh token to keep login state"
        },
        "expiresIn": {
          "type": "integer",
          "format": "int32",
          "title": "Expires time in seconds"
        },
        "redirect": {
          "type": "string"
        }
      }
    },
    "v1WebLoginAuthRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "required": [
            "username"
          ]
        },
        "password": {
          "type": "string",
          "required": [
            "password"
          ]
        },
        "remember": {
          "type": "boolean"
        },
        "challenge": {
          "type": "string"
        },
        "reject": {
          "type": "boolean"
        }
      },
      "required": [
        "username",
        "password"
      ]
    }
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "description": "Authentication token, prefixed by Bearer: Bearer \u003ctoken\u003e",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}